monitor/requirements.txt
flask
flask-apscheduler
flask-sqlalchemy
gunicorn
plumbum
psycopg[binary,pool]
